requirements.txt
requests
dnspython
tqdm
